updated start methods
